diff_odometry: change msg type from UInt8 to Float64
jetson_i2c: added launch file, commented publishing of ultrasonic
sensores

watchdog_startup: added two launch  files for quick and easy startup,
one for physical robot nodes and one for simulation

+++ diff_odometry/include/diff_odometry/diff_odometry.h
@@ -4,22 +4,22 @@
 #define DIFF_ODOM_H
 
     #include <ros/ros.h>
-    #include <std_msgs/UInt8.h>
+    #include <std_msgs/Float64.h>
     #include <tf/transform_broadcaster.h>
     #include <nav_msgs/Odometry.h>
     #include <iostream>
 
     void calculateOdometry(double delta_time);
     void initializeGlovalVariables();
-    void spd_msgCallback_left(const std_msgs::UInt8& msg);
-    void spd_msgCallback_right(const std_msgs::UInt8& msg);
+    void spd_msgCallback_left(const std_msgs::Float64& msg);
+    void spd_msgCallback_right(const std_msgs::Float64& msg);
 
     const double _WHEEL_RADIUS = 0.0509; // meters
     const double _WHEEL_PERIMETER = 2 * 3.14159265 * _WHEEL_RADIUS;
     const double _SEPARATION_BETWEEN_WHEELS = 0.225; // meters
 
-    extern std_msgs::UInt8 speed_left;
-    extern std_msgs::UInt8 speed_right;
+    extern std_msgs::Float64 speed_left;
+    extern std_msgs::Float64 speed_right;
     //////////////////////////////////////////////////////////////////////////// ODOM SETUP
     // Positions
     extern double x;    // Robot starts at the origin of the "odom" coordinate frame initially
